+++ feature/main/src/main/java/kr/co/main/calendar/screen/addDiaryScreen/diary_input/DiaryOverviewInput.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import kr.co.main.R
-import kr.co.main.calendar.ui.common.CalendarUnderLineTextField
+import kr.co.main.calendar.common.CalendarUnderLineTextField
 import kr.co.ui.theme.Paddings
 import kr.co.ui.theme.typo
