+++ feature/main/src/main/java/kr/co/main/calendar/screen/calendarScreen/scheduleTab/ScheduleTab.kt
@@ -25,13 +25,13 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import kr.co.main.R
+import kr.co.main.calendar.common.CalendarCategoryIndicator
+import kr.co.main.calendar.common.innerCalendar.InnerCalendar
 import kr.co.main.model.calendar.CropModel
 import kr.co.main.model.calendar.FarmWorkModel
 import kr.co.main.model.calendar.HolidayModel
 import kr.co.main.model.calendar.ScheduleModel
 import kr.co.main.model.calendar.filterAndSortHolidays
-import kr.co.main.calendar.ui.common.CalendarCategoryIndicator
-import kr.co.main.calendar.ui.common.innerCalendar.InnerCalendar
 import kr.co.ui.theme.Paddings
 import kr.co.ui.theme.colors
 import kr.co.ui.theme.typo
