+++ feature/main/src/main/java/kr/co/main/navigation/MainNavigation.kt
@@ -8,10 +8,10 @@ import kr.co.main.MainRoute
 import kr.co.main.accountbook.content.AccountBookContentRoute
 import kr.co.main.accountbook.main.AccountBookRoute
 import kr.co.main.accountbook.register.AccountBookRegisterRoute
-import kr.co.main.calendar.ui.screen.addDiaryScreen.AddDiaryRoute
-import kr.co.main.calendar.ui.screen.addScheduleScreen.AddScheduleRoute
-import kr.co.main.calendar.ui.screen.calendarScreen.CalendarRoute
-import kr.co.main.calendar.ui.screen.searchDiaryScreen.SearchDiaryRoute
+import kr.co.main.calendar.screen.addDiaryScreen.AddDiaryRoute
+import kr.co.main.calendar.screen.addScheduleScreen.AddScheduleRoute
+import kr.co.main.calendar.screen.calendarScreen.CalendarRoute
+import kr.co.main.calendar.screen.searchDiaryScreen.SearchDiaryRoute
 import kr.co.main.community.BulletinDetailRoute
 import kr.co.main.community.BulletinWritingRoute
 import kr.co.main.community.CommunityRoute
